requirements.txt
fastapi==0.111.1
pony==0.7.17
strawberry-graphql==0.237.1
uvicorn==0.30.3
